@@ -1,14 +1,14 @@
-package com.abemart.wroupchat;
+package com.abemart.wroup;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import com.abemart.wroup.common.WroupDevice;
 import com.abemart.wroup.common.messages.MessageWrapper;
 
@@ -16,15 +16,6 @@ import java.util.List;
 
 
 public class ChatAdapter extends ArrayAdapter<MessageWrapper> {
-
-    private static class ChatAdapterHolder {
-        TextView txtViewUsername;
-        TextView txtViewMessage;
-    }
-
-    private static class ChatAdapterOwnerHolder {
-        TextView txtViewMessage;
-    }
 
     private WroupDevice currentDevice;
 
@@ -46,31 +37,47 @@ public class ChatAdapter extends ArrayAdapter<MessageWrapper> {
                 convertView = inflater.inflate(R.layout.adapter_chat_owner, parent, false);
 
                 chatAdapterOwnerHolder = new ChatAdapterOwnerHolder();
-                chatAdapterOwnerHolder.txtViewMessage = (TextView) convertView.findViewById(R.id.text_view_message_owner);
+                chatAdapterOwnerHolder.txtViewMessageId = convertView.findViewById(R.id.text_view_message_Id_owner);
+                chatAdapterOwnerHolder.txtViewMessage = convertView.findViewById(R.id.text_view_message_owner);
                 convertView.setTag(chatAdapterOwnerHolder);
             } else {
                 chatAdapterOwnerHolder = (ChatAdapterOwnerHolder) convertView.getTag();
             }
 
-            chatAdapterOwnerHolder.txtViewMessage.setText(message.getMessage());
+            chatAdapterOwnerHolder.txtViewMessageId.setText(String.valueOf(message.getPipelineObject().getId()));
+            chatAdapterOwnerHolder.txtViewMessage.setText(message.getPipelineObject().getName());
         } else {
             ChatAdapterHolder chatAdapterHolder;
             if (convertView == null) {
                 LayoutInflater inflater = LayoutInflater.from(getContext());
-                convertView = inflater.inflate(R.layout.adapter_chat, parent, false);
+                convertView = inflater.inflate(R.layout.list_item_object, parent, false);
 
                 chatAdapterHolder = new ChatAdapterHolder();
-                chatAdapterHolder.txtViewUsername = (TextView) convertView.findViewById(R.id.text_view_username);
-                chatAdapterHolder.txtViewMessage = (TextView) convertView.findViewById(R.id.text_view_message);
+                chatAdapterHolder.txtViewUsername = convertView.findViewById(R.id.text_device_username);
+                chatAdapterHolder.txtViewMessageId = convertView.findViewById(R.id.text_object_id);
+                chatAdapterHolder.txtViewMessage = convertView.findViewById(R.id.text_object_name);
                 convertView.setTag(chatAdapterHolder);
             } else {
                 chatAdapterHolder = (ChatAdapterHolder) convertView.getTag();
             }
 
+            Pipeline pipelineInstance = message.getPipelineObject();
             chatAdapterHolder.txtViewUsername.setText(message.getWroupDevice().getDeviceName());
-            chatAdapterHolder.txtViewMessage.setText(message.getMessage());
+            chatAdapterHolder.txtViewMessageId.setText(String.valueOf(pipelineInstance.getId()));
+            chatAdapterHolder.txtViewMessage.setText(pipelineInstance.getName());
         }
 
         return convertView;
+    }
+
+    private static class ChatAdapterHolder {
+        TextView txtViewUsername;
+        TextView txtViewMessageId;
+        TextView txtViewMessage;
+    }
+
+    private static class ChatAdapterOwnerHolder {
+        TextView txtViewMessageId;
+        TextView txtViewMessage;
     }
 }
