@@ -1,4 +1,4 @@
-package com.abemart.wroupchat;
+package com.abemart.wroup;
 
 import android.app.ProgressDialog;
 import android.content.DialogInterface;
@@ -6,12 +6,13 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.wifi.p2p.WifiP2pManager;
 import android.os.Bundle;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.abemart.wroup.client.WroupClient;
 import com.abemart.wroup.common.WiFiDirectBroadcastReceiver;
@@ -141,7 +142,7 @@ public class MainActivity extends AppCompatActivity implements GroupCreationDial
                 progressDialog.dismiss();
 
                 if (serviceDevices.isEmpty()) {
-                    Toast.makeText(getApplicationContext(), getString(R.string.toast_not_found_groups),Toast.LENGTH_LONG).show();
+                    Toast.makeText(getApplicationContext(), getString(R.string.toast_not_found_groups), Toast.LENGTH_LONG).show();
                 } else {
                     showPickGroupDialog(serviceDevices);
                 }
